requirements.txt
# Core ML and Deep Learning
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
bitsandbytes>=0.39.0

# Data Processing and Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Web Scraping and Data Collection
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.10.0
fastf1>=3.1.0
kaggle>=1.5.0

# Visualization and Plotting
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Experiment Tracking and Logging
wandb>=0.15.0
mlflow>=2.5.0
tensorboard>=2.13.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=6.5.0

# API and Web Services
fastapi>=0.100.0
uvicorn>=0.23.0
openai>=0.27.0
anthropic>=0.3.0

# Data Storage and Cloud
google-cloud-storage>=2.10.0
boto3>=1.28.0
azure-storage-blob>=12.17.0

# Configuration and Environment
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0

# Testing and Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Utilities
tqdm>=4.65.0
rich>=13.4.0
click>=8.1.0
loguru>=0.7.0

# Text Processing
nltk>=3.8.0
spacy>=3.6.0
tokenizers>=0.13.0

# Evaluation Metrics
rouge-score>=0.1.2
sacrebleu>=2.3.0
bert-score>=0.3.13

# Model Serving and Deployment
gradio>=3.40.0
streamlit>=1.25.0
docker>=6.1.0

# Development Tools
pre-commit>=3.3.0
jupyterlab>=4.0.0
ipywidgets>=8.0.0
